+++ .github/workflows/test.yaml
@@ -47,3 +47,12 @@ jobs:
       - run: yarn prepare
 
       - run: yarn typecheck
+
+  build:
+    runs-on: ubuntu-latest
+    steps:
+      - uses: actions/checkout@v3
+      - uses: ./.github/actions/setup
+      - run: yarn prepare
+
+      - run: yarn build
